Change error import locations
