+++ lib/permissions/pages/actions/upsert-permission.ts
@@ -2,7 +2,8 @@ import { PagePermission, PagePermissionLevel, Prisma } from '@prisma/client';
 import { prisma } from 'db';
 import { resolveChildPagesAsFlatList, resolveParentPages } from 'lib/pages/server';
 import { isTruthy } from 'lib/utilities/types';
-import { CannotInheritOutsideTreeError, InvalidPermissionGranteeError, InvalidPermissionLevelError, PermissionNotFoundError, SelfInheritancePermissionError } from '../errors';
+import { InvalidPermissionGranteeError } from 'lib/permissions/errors';
+import { CannotInheritOutsideTreeError, InvalidPermissionLevelError, PermissionNotFoundError, SelfInheritancePermissionError } from '../errors';
 import { IPagePermissionToCreate, IPagePermissionWithSource } from '../page-permission-interfaces';
 import { checkParentForSamePermission } from './check-parent-for-same-permission';
